Organize code
- Remove event.js file and put all its events in the application file
- Create custom event to reset materialize components
