+++ index.html
@@ -30,8 +30,6 @@
     <script type="text/javascript" src="js/materialize.min.js" ></script>
     <script type="text/javascript" src="js/handlebars.runtime-v4.1.2.js"></script>
     <script type="text/javascript" src="js/templates_compiled.js"></script>
-    <script type="text/javascript" src="js/book.js"></script>
-    <script type="text/javascript" src="js/events.js"></script>
-    <script type="text/javascript" src="js/application.js"></script>
+    <script type="module" src="js/application.js"></script>
   </body>
 </html>
